requirements.txt
# Core Flask dependencies
Flask==3.0.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-CORS==4.0.0
Werkzeug==3.0.1

# Database
PyMySQL==1.1.0
SQLAlchemy==2.0.36

# Password hashing
bcrypt==4.1.2

# Forms and validation
WTForms==3.1.1
email-validator==2.1.0
phonenumbers==8.13.27

# CSV and Excel handling
openpyxl==3.1.2
xlsxwriter==3.1.9

# PDF generation
reportlab==4.0.7
PyPDF2==3.0.1

# Environment variables
python-dotenv==1.0.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
requests==2.31.0

# Development
python-json-logger==2.0.7
